refactor: improve code readability by moving components for better co-location
