+++ apps/food-planner-client/src/components/ScheduleView/components/ScheduleRecipeSelect/types.ts
@@ -1,6 +1,6 @@
 import type { Control, FieldValues, Path } from "react-hook-form";
 
-export type TRecipeSelectProps<T extends FieldValues> = {
+export type TScheduleRecipeSelectProps<T extends FieldValues> = {
   control: Control<T>;
   name: Path<T>;
 };
