+++ apps/food-planner-client/src/components/ShoppingList/components/ShoppingListItem/index.tsx
