+++ apps/food-planner-client/src/layouts/MainLayout/components/AppSidebar/const.ts
