+++ apps/food-planner-client/src/layouts/MainLayout/components/MobileSidebarHeader/index.tsx
